Add unit tests for WeightLogServiceImpl, refactor methods and add ApplicationBeanConfiguration class with beans for mappers

+++ src/main/java/com/hyuseinlesho/powerlog/service/impl/WeightLogServiceImpl.java
@@ -1,11 +1,11 @@
 package com.hyuseinlesho.powerlog.service.impl;
 
+import com.hyuseinlesho.powerlog.exception.WeightLogNotFoundException;
 import com.hyuseinlesho.powerlog.mapper.WeightLogMapper;
 import com.hyuseinlesho.powerlog.model.dto.CreateWeightLogDto;
 import com.hyuseinlesho.powerlog.model.dto.UpdateWeightLogDto;
 import com.hyuseinlesho.powerlog.model.dto.WeightLogDto;
 import com.hyuseinlesho.powerlog.model.dto.WeightLogGraphDto;
-import com.hyuseinlesho.powerlog.model.entity.UserEntity;
 import com.hyuseinlesho.powerlog.model.entity.WeightLog;
 import com.hyuseinlesho.powerlog.repository.WeightLogRepository;
 import com.hyuseinlesho.powerlog.service.UserService;
@@ -18,19 +18,19 @@ import java.util.List;
 @Service
 public class WeightLogServiceImpl implements WeightLogService {
     private final WeightLogRepository weightLogRepository;
+    private final WeightLogMapper weightLogMapper;
     private final UserService userService;
 
-    public WeightLogServiceImpl(WeightLogRepository weightLogRepository, UserService userService) {
+    public WeightLogServiceImpl(WeightLogRepository weightLogRepository, WeightLogMapper weightLogMapper, UserService userService) {
         this.weightLogRepository = weightLogRepository;
+        this.weightLogMapper = weightLogMapper;
         this.userService = userService;
     }
 
     @Override
     public WeightLog createWeightLog(CreateWeightLogDto weightLogDto) {
-        UserEntity currentUser = userService.getCurrentUser();
-
-        WeightLog weightLog = WeightLogMapper.INSTANCE.mapToWeightLog(weightLogDto);
-        weightLog.setUser(currentUser);
+        WeightLog weightLog = weightLogMapper.mapToWeightLog(weightLogDto);
+        weightLog.setUser(userService.getCurrentUser());
 
         return weightLogRepository.save(weightLog);
     }
@@ -40,13 +40,14 @@ public class WeightLogServiceImpl implements WeightLogService {
         List<WeightLog> weightLogs = weightLogRepository.findAllByUser(userService.getCurrentUser());
 
         return weightLogs.stream()
-                .map(WeightLogMapper.INSTANCE::mapToWeightLogDto)
+                .map(weightLogMapper::mapToWeightLogDto)
                 .toList();
     }
 
     @Override
     public WeightLog updateWeightLog(Long id, UpdateWeightLogDto weightLogDto) {
-        WeightLog weightLog = weightLogRepository.findById(id).get();
+        WeightLog weightLog = weightLogRepository.findById(id)
+                .orElseThrow(() -> new WeightLogNotFoundException("WeightLog not found for id: " + id));
 
         weightLog.setWeight(weightLogDto.getWeight());
         weightLog.setDate(weightLogDto.getDate());
@@ -58,21 +59,26 @@ public class WeightLogServiceImpl implements WeightLogService {
 
     @Override
     public void deleteWeightLog(Long id) {
+        if (!weightLogRepository.existsById(id)) {
+            throw new WeightLogNotFoundException("WeightLog not found for id: " + id);
+        }
+
         weightLogRepository.deleteById(id);
     }
 
     @Override
     public List<WeightLogGraphDto> getWeightLogs() {
-        return weightLogRepository.findByUserOrderByDateAsc(userService.getCurrentUser())
+        return weightLogRepository.findAllByUserOrderByDateAsc(userService.getCurrentUser())
                 .stream()
-                .map(WeightLogMapper.INSTANCE::mapToWeightLogGraphDto)
+                .map(weightLogMapper::mapToWeightLogGraphDto)
                 .toList();
     }
 
     @Override
     public List<WeightLogGraphDto> getWeightLogsBetweenDates(LocalDate startDate, LocalDate endDate) {
-        return weightLogRepository.findByDateBetween(startDate, endDate).stream()
-                .map(WeightLogMapper.INSTANCE::mapToWeightLogGraphDto)
+        return weightLogRepository.findAllByDateBetweenAndUser(startDate, endDate, userService.getCurrentUser())
+                .stream()
+                .map(weightLogMapper::mapToWeightLogGraphDto)
                 .toList();
     }
 }
